+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute_sediment/fm_adjust_bedload.f90
@@ -73,10 +73,10 @@ contains
       logical :: di50spatial
       integer :: l, Lf, k1, k2, lb, lt
 
-      double precision :: di50, phi, tphi, sbedm, depth, dzdp, dzds, bagnol, alfas
-      double precision :: delta, dmloc, ftheta, hidexploc, shield, sina, cosa, tnorm, frc, fixf
-      double precision :: sbedn, sbedt, tratio, sbedcorr, fnorm, ust2avg, slp, avflux
-      double precision :: eps = 1.0d-6
+      real(kind=dp) :: di50, phi, tphi, sbedm, depth, dzdp, dzds, bagnol, alfas
+      real(kind=dp) :: delta, dmloc, ftheta, hidexploc, shield, sina, cosa, tnorm, frc, fixf
+      real(kind=dp) :: sbedn, sbedt, tratio, sbedcorr, fnorm, ust2avg, slp, avflux
+      real(kind=dp) :: eps = 1.0d-6
       !
    !! executable statements -------------------------------------------------------
       !
