+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_utils/filez.F90
@@ -530,7 +530,7 @@ subroutine zoekdouble(minp, key, val, ja)
    implicit none
    integer, intent(in) :: minp !< File pointer
    character(*), intent(in) :: key
-   double precision, intent(out) :: val !<
+   real(kind=dp), intent(out) :: val !<
    integer, intent(out) :: ja !< Whether key was found or not.
 
    character(len=255) :: rec, key2
